+++ main/api/user-settings/index.tsx
@@ -1,34 +1,33 @@
 import { nativeTheme } from "electron";
 import { z } from "zod";
 
-import { PartialUserSettingsSchema } from "./userSettings";
-import { grabbyUserSettings, settingsKeys } from "./userSettingsAlt";
+import { getUserSettings, settingsZodSchema } from "./userSettings";
 import { t } from "../trpc";
 
 export const userSettingsRouter = t.router({
   getUserSetting: t.procedure
     .input(
       z.object({
-        key: z.enum(settingsKeys),
+        key: settingsZodSchema.keyof(),
       }),
     )
     .query(async (opts) => {
-      const settings = await grabbyUserSettings();
+      const settings = await getUserSettings();
       return settings.get(opts.input.key);
     }),
   getUserSettings: t.procedure.query(async () => {
-    const settings = await grabbyUserSettings();
+    const settings = await getUserSettings();
     return settings;
   }),
   setUserSettings: t.procedure
-    .input(PartialUserSettingsSchema)
+    .input(settingsZodSchema.partial())
     .mutation(async (opts) => {
       if (Object.hasOwn(opts.input, "theme") && opts.input.theme) {
         const themeValue = opts.input.theme;
         nativeTheme.themeSource = themeValue;
       }
 
-      const settings = await grabbyUserSettings();
+      const settings = await getUserSettings();
       settings.set(opts.input);
     }),
 });
