+++ main/api/user-settings/userSettings.ts
@@ -1,55 +1,46 @@
-import { DEFAULT_DATA_DIR, KeyStore, NodeFileProvider } from "@ironfish/sdk";
-import { app } from "electron";
+import { DEFAULT_DATA_DIR, NodeFileProvider } from "@ironfish/sdk";
+import Store, { Schema } from "electron-store";
 import { z } from "zod";
 
-export type UserSettings = {
-  enabled: boolean;
-  dataDir: string;
-  theme: "light" | "dark" | "system";
-  locale: string | null;
+const STORE_NAME = "user-settings";
+
+export const settingsZodSchema = z.object({
+  dataDir: z.string(),
+  theme: z.enum(["light", "dark", "system"]),
+  locale: z.string().nullable(),
+});
+
+export type SchemaDefinition = z.infer<typeof settingsZodSchema>;
+
+export const settingsSchema: Schema<SchemaDefinition> = {
+  dataDir: {
+    type: "string",
+  },
+  theme: {
+    enum: ["light", "dark", "system"],
+  },
+  locale: {
+    type: "string",
+  },
 };
 
-export const PartialUserSettingsSchema: z.ZodType<Partial<UserSettings>> = z
-  .object({
-    enabled: z.boolean(),
-    dataDir: z.string(),
-    theme: z.enum(["light", "dark", "system"]),
-    locale: z.string().nullable(),
-  })
-  .partial();
+let settingsStore: Store<SchemaDefinition>;
 
-export class UserSettingsStore extends KeyStore<UserSettings> {
-  constructor(options: {
-    fileSystem: NodeFileProvider;
-    dataDir: string;
-    fileName: string;
-    defaults: UserSettings;
-  }) {
-    super(
-      options.fileSystem,
-      options.fileName,
-      options.defaults,
-      options.dataDir,
-    );
-  }
-}
+export async function grabbyUserSettings() {
+  if (settingsStore) return settingsStore;
 
-export async function loadUserSettings(): Promise<UserSettingsStore> {
   const fileSystem = new NodeFileProvider();
   await fileSystem.init();
 
-  const userSettings = new UserSettingsStore({
-    fileSystem,
-    dataDir: app.getPath("userData"),
-    fileName: "settings.json",
+  settingsStore = new Store({
+    schema: settingsSchema,
+    name: STORE_NAME,
     defaults: {
       dataDir: fileSystem.resolve(DEFAULT_DATA_DIR),
-      enabled: false,
       theme: "system",
       locale: null,
     },
   });
 
-  await userSettings.load();
-  return userSettings;
+  return settingsStore;
 }
